requirements.txt
Django==4.2.11
django-admin-rangefilter==0.12.4
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==6.7.1
Faker==24.8.0
